requirements.txt
# Google APIs for YouTube
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
google-auth>=2.15.0

# Facebook/Instagram APIs (TODO: add when implementing)
# facebook-sdk

# TikTok APIs (TODO: add when implementing) 
# tiktok-api

# General dependencies
requests>=2.28.0
urllib3>=1.26.0
python-dateutil>=2.8.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0